Update pwdhash-safari.js
there is a bug when you press i.e. shift to insert an uppercase letter but it adds three characters (key code for shift obviously) plus the uppercase letter generating a flash hash of course.
removing this fixed it for me.
unfortunately I do not own a certificate so I can't build it :(
could you rebuild the extension?